Separate files for dialog callbacks.
