+++ src/engine/scripts/declarations/dialogs/dialogs/index.ts
@@ -1,2 +1,4 @@
 require("@/engine/scripts/declarations/dialogs/dialogs/dialogs_generic");
+require("@/engine/scripts/declarations/dialogs/dialogs/dialogs_object");
 require("@/engine/scripts/declarations/dialogs/dialogs/dialogs_quest");
+require("@/engine/scripts/declarations/dialogs/dialogs/dialogs_world");
